Remove threadpool usage.
Since the thread pool threads a long lived, we're not able to borrow
atomics into them, which means we instead have to wrap our atomics in
Arc<>s. We can use thread scopes instead.

In a real game, we'll spawn the threads once per turn which is small
compared the to several seconds which we'll be seaching for.