Implement mvv_lva
We already generate captures in order of least valuable attack so this
doesn't actually gain us much. Benchmarks say it there's no significant
difference either way. The cost of creating the Vec roughly balances out
the couple of branches that we need to search.

Howver, it opens the door for more efficient, non iterative capture
generation since we need to collect and sort the moves anyways.